backend/requirements.txt
fastapi>=0.104.1,<2.0.0
uvicorn>=0.23.2,<0.24.0
python-multipart>=0.0.6,<0.1.0
sqlalchemy>=2.0.23,<3.0.0
aiosqlite>=0.19.0,<0.20.0
pydantic>=2.4.2,<3.0.0
langchain>=0.0.335,<0.1.0
langchain-openai>=0.0.2,<0.1.0
langchain-google-genai>=0.0.5,<0.1.0
openai>=1.3.0,<2.0.0
google-api-python-client>=2.107.0,<3.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
google-auth-httplib2>=0.1.1,<0.2.0
PyPDF2>=3.0.1,<4.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.25.1,<0.26.0